.Projects/screener/Data/requirements.txt
colorama
opencv-python
numpy
pillow